Add the Calibration Test window and enable recording from it
