+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/Editor/CalibrationTestWindow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9f8b117576190504d94626a1805694fa
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
