+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/Editor/CompositorWindowBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 262977c55f5a63344be87a92e712e63e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
